updated tests on auth, book, server, userbooks, and created test-object file
