+++ test/test-objects.js
@@ -1,0 +1,56 @@
+module.exports = {
+	bookObject: {
+		googleId: "qwoldmcdfiom123103",
+		title: "Chantra Swandie",
+		authors: "McWorld",
+		publisher: "Penguin",
+		publishedDate: "2/21/2020",
+		description: "The end of the book",
+		isbn10: "12345678911234567891",
+		isbn13: "12345678911234567891234",
+		pageCount: 210,
+		categories: "swenad",
+		thumbnail: "image.png",
+		smallThumbnail: "small-img.png",
+		language: "english",
+		webReaderLink: "testLink",
+		textSnippet: "testSnippet",
+		isEbook: true,
+		averageRating: 4
+	},
+
+	otherBook: {
+		googleId: "qwertyomsname",
+		title: "Lander McPherson",
+		authors: "Civil Mary",
+		publisher: "Top hat",
+		publishedDate: "4/2/1931",
+		description: "The begining of the book",
+		isbn10: "0293129582812931832914",
+		isbn13: "90w8q9weqw9eq0w9e0w9eq9",
+		pageCount: 100,
+		categories: "mapry",
+		thumbnail: "image.png",
+		smallThumbnail: "small-img.png",
+		language: "english",
+		webReaderLink: "testLink",
+		textSnippet: "testSnippet",
+		isEbook: false,
+		averageRating: 4
+	},
+
+	badBookObject : {
+		type: "Movie",
+		content: "Im not a book"
+	},
+
+	shelfObj: {
+		shelfName: "Test shelf",
+    	isPrivate: false
+	},
+	
+	shelfWithBook: {
+	    bookId: 1,
+    	shelfId: 1	
+	}
+};
