+++ test/user-books-on-a-shelf.test.js
@@ -1,8 +1,11 @@
 const server = require("../api/server.js");
 const request = require("supertest");
 const db = require("../database/db-config.js");
-
+const TestObject = require("./test-objects.js");
 const knexCleaner = require('knex-cleaner');
+
+const shelfObj = TestObject.shelfObj;
+const bookObject = TestObject.bookObject
 
 var options = {
 	mode: 'truncate',
@@ -12,59 +15,6 @@ var options = {
 };
 
 describe("user-books-on-a-shelf", function() {
-	
-const body = {
-    book: {
-            googleId: "8888ATEATE",
-            title: "Don't Trust A Bank",
-            authors: "Edd Stark",
-            publisher: "Ice&Fire",
-            publishedDate: "2/21/2020",
-            description: "Something or another",
-            isbn10: "12345678911234567891",
-            isbn13: "12345678911234567891234",
-            pageCount: 210,
-            categories: "swenad",
-            thumbnail: "image.png",
-            smallThumbnail: "small-img.png",
-            language: "english",
-            webReaderLink: "testLink",
-            textSnippet: "testSnippet",
-            isEbook: true,
-            averageRating: 5
-        },
-        readingStatus: 1,
-        favorite: true
-    }
-
-    const bookObject2 = {
-            googleId: "999NINNER",
-            title: "In those genes",
-            authors: "Emmit Smith",
-            publisher: "Austin Co.",
-            publishedDate: "8/2/1980",
-            description: "The birth of dirt",
-            isbn10: "12345678911234567891",
-            isbn13: "12345678911234567891234",
-            pageCount: 210,
-            categories: "swenad",
-            thumbnail: "image.png",
-            smallThumbnail: "small-img.png",
-            language: "english",
-            webReaderLink: "testLink",
-            textSnippet: "testSnippet",
-            isEbook: true,
-            averageRating: 4
-        };
-        const shelfObj = {
-            shelfName: "Test shelf",
-            isPrivate: false
-        }
-
-        const bookObj = {
-            bookId: 1,
-            shelfId: 1
-          }
 
 	// MARK: -- helper function to grab cookie
 	function promisedCookie(user) {
@@ -90,63 +40,77 @@ const body = {
 				password: "seedpassword"
 			}).then(res => {
 				const cookie = res.headers["set-cookie"]
-				return request(server)
-					.post("/api/books")
-					.send(bookObject2)
-					.set("cookie", cookie)
-					.then(res => {
-						return request(server)
-							.post('/api/1/library')
-							.send({ book: bookObject2, readingStatus: 2 })
-                            .set("cookie", cookie)
-                            .then(res => {
-                                return request(server)
-                                    .post('/api/shelves/user/1')
-                                    .send(shelfObj)
-                                    .set("cookie", cookie)
-                                    .then(res => {
-                                        return request(server)
-                                            .post('/api/booksonshelf/shelves/1')
-                                            .send({ book: bookObject2, readingStatus: 2 })
-                                            .set("cookie", cookie)
-                                                                    
-                                    })                                      
-                            })                        
-					})
-			})
+                return request(server)
+                    .post('/api/shelves/user/1')
+                    .send(shelfObj)
+                    .set("cookie", cookie)
+                    .then(res => {
+                        return request(server)
+                            .post('/api/booksonshelf/shelves/1')
+                            .send({ book: bookObject, readingStatus: 2 })
+                            .set("cookie", cookie)                             
+					   })
+			     })
 	});
     
     describe("GET user books on shelf user shelfId", function() {
-		it("GET res is shelf.Id", function() {
+		it("GET /booksonshelf/shelves/:shelfId", function() {
 			return promisedCookie({ emailAddress: "seedemail", password: "seedpassword" }).then(cookie => {
 				const req = request(server)
                     .get("/api/booksonshelf/shelves/1")
-					.send({bookId: 1})
+					.send({ bookId: 1 })
 					.set("cookie", cookie)
 					.then(res => {
-                        
-						expect(res.body[0].shelfId).toBe(1);
+						expect(res.status).toBe(200);
 					});
                 return req;
             
 			});
         });
 
-        it("GET res is user.Id", function() {
+        it("GET /booksonshelf/shelves/allbooks/:shelfId", function() {
             return promisedCookie({ emailAddress: "seedemail", password: "seedpassword" }).then(cookie => {
                 const req = request(server)
-                    .get("/api/booksonshelf/shelves/1")
-                    .send({bookId: 1})
+                    .get("/api/booksonshelf/shelves/allbooks/1")
+                    .send({ userId: 1 })
                     .set("cookie", cookie)
                     .then(res => {
-                        expect(res.body[0].userId).toBe(1);
+                        expect(res.status).toBe(200);
                     });
                 return req;
             
             });
         });
-        
+
+        it("GET /booksonshelf/user/:userId", function() {
+            return promisedCookie({ emailAddress: "seedemail", password: "seedpassword" }).then(cookie => {
+                const req = request(server)
+                    .get("/api/booksonshelf/user/1")
+                    .set("cookie", cookie)
+                    .then(res => {
+                        expect(res.body[0].shelfName).toBe('Test shelf');
+                    });
+                return req;
+            
+            });
+        });
     });
+
+    describe("PUT user books on shelf", function() {
+        it("PUT /booksonshelf/shelves/:shelfId", function() {
+            return promisedCookie({ emailAddress: "seedemail", password: "seedpassword" }).then(cookie => {
+                const req = request(server)
+                    .put("/api/booksonshelf/shelves/1")
+                    .send({ bookId: 1, newShelfId: 2 })
+                    .set("cookie", cookie)
+                    .then(res => {
+                        expect(res.body.message).toBe("error in moving book to shelf");
+                    });
+                return req;
+            
+            });
+        });
+    })
     
     describe("DELETE from shelf", function() {
         it("Delete book from shelf", function () {
